Revert "UI 동기화를 위한 작업들"
This reverts commit 37ab907474b6cdd4382dfce6b2d9d0df035b7c02.

+++ Project/2019180016/IOCP-Test/ClientTest/Source/ClientTest/ClientSocketMgr.cpp
@@ -182,10 +182,10 @@ uint32 ClientSocketMgr::Run()
 		}
 
 		std::stringstream RecvStream(m_sRecvBuffer);
-		BYTE OP;
+		int OP;
 		RecvStream >> OP;
 
-		memcpy(&OP, m_sRecvBuffer, sizeof(BYTE));
+		memcpy(&OP, m_sRecvBuffer, sizeof(int));
 
 		switch (OP)
 		{
